Improve automatic detaching / deletion of relations (#156)
Adds support for automatic detaching / deletion of relationships using pivot tables.

Related: wintercms/winter#386, wintercms/winter#972.
Documented by https://github.com/wintercms/docs/pull/159.

+++ src/Database/Traits/SoftDelete.php
@@ -115,21 +115,27 @@ trait SoftDelete
         $definitions = $this->getRelationDefinitions();
         foreach ($definitions as $type => $relations) {
             foreach ($relations as $name => $options) {
-                if (!array_get($options, 'softDelete', false)) {
-                    continue;
-                }
-
                 if (!$relation = $this->{$name}) {
                     continue;
                 }
-
-                if ($relation instanceof EloquentModel) {
-                    $relation->delete();
+                if (!array_get($options, 'softDelete', false)) {
+                    continue;
                 }
-                elseif ($relation instanceof CollectionBase) {
-                    $relation->each(function ($model) {
-                        $model->delete();
-                    });
+                if (in_array($type, ['belongsToMany', 'morphToMany', 'morphedByMany'])) {
+                    // relations using pivot table
+                    $value = $this->fromDateTime($this->freshTimestamp());
+                    $this->updatePivotDeletedAtColumn($name, $options, $value);
+                } elseif (in_array($type, ['belongsTo', 'hasOneThrough', 'hasManyThrough', 'morphTo'])) {
+                    // the model does not own the related record, we should not remove it.
+                    continue;
+                } elseif (in_array($type, ['attachOne', 'attachMany', 'hasOne', 'hasMany', 'morphOne', 'morphMany'])) {
+                    if ($relation instanceof EloquentModel) {
+                        $relation->delete();
+                    } elseif ($relation instanceof CollectionBase) {
+                        $relation->each(function ($model) {
+                            $model->delete();
+                        });
+                    }
                 }
             }
         }
@@ -178,6 +184,19 @@ trait SoftDelete
     }
 
     /**
+     * Update relation pivot table deleted_at column
+     */
+    protected function updatePivotDeletedAtColumn(string $relationName, array $options, string|null $value)
+    {
+        // get deletedAtColumn from the relation options, otherwise use default
+        $deletedAtColumn = array_get($options, 'deletedAtColumn', 'deleted_at');
+
+        $this->{$relationName}()->newPivotQuery()->update([
+            $deletedAtColumn => $value,
+        ]);
+    }
+
+    /**
      * Locates relations with softDelete flag and cascades the restore event.
      *
      * @return void
@@ -191,18 +210,22 @@ trait SoftDelete
                     continue;
                 }
 
-                $relation = $this->{$name}()->onlyTrashed()->getResults();
-                if (!$relation) {
-                    continue;
-                }
+                if (in_array($type, ['belongsToMany', 'morphToMany', 'morphedByMany'])) {
+                    // relations using pivot table
+                    $this->updatePivotDeletedAtColumn($name, $options, null);
+                } else {
+                    $relation = $this->{$name}()->onlyTrashed()->getResults();
+                    if (!$relation) {
+                        continue;
+                    }
 
-                if ($relation instanceof EloquentModel) {
-                    $relation->restore();
-                }
-                elseif ($relation instanceof CollectionBase) {
-                    $relation->each(function ($model) {
-                        $model->restore();
-                    });
+                    if ($relation instanceof EloquentModel) {
+                        $relation->restore();
+                    } elseif ($relation instanceof CollectionBase) {
+                        $relation->each(function ($model) {
+                            $model->restore();
+                        });
+                    }
                 }
             }
         }
